llm-server/requirements.txt
annotated-types
cachetools
certifi
cffi
charset-normalizer
colorama
cryptography
distlib
dotenv
filelock
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
pdfminer.six
pdfplumber
pillow
platformdirs
plum-dispatch
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyparsing
pypdfium2
python-dotenv
requests
rsa
Spire.Doc
tqdm
typing_extensions
uritemplate
urllib3
virtualenv
websockets
